[YS-257] refactor: 진행 방식 필터링 컴포넌트명 수정
