+++ src/app/home/components/PostContainer/FilterContainer/FilterContainer.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-export const filterLayout = style({
+export const filterContainerLayout = style({
   display: 'flex',
   gap: '0.8rem',
   height: '3.2rem',
