+++ src/app/home/components/PostContainer/FilterContainer/FilterContainer.tsx
@@ -1,17 +1,32 @@
-import { filterLayout } from './FilterContainer.css';
+import { filterContainerLayout } from './FilterContainer.css';
 import AreaFilter from '../AreaFilter/AreaFilter';
 import ContactTargetFilter from '../ContactTargetPopover/ContactTargetFilter';
-import ProgressMethodFilter from '../ProgressMethodFilter/ProgressMethodFilter';
+import MatchTypeFilter from '../MatchTypeFilter/MatchTypeFilter';
+
+import { ExperimentPostListFilters } from '@/apis/post';
 
 interface FilterContainerProps {
+  filters: ExperimentPostListFilters;
   handleFilterChange: (key: string, value: string | number) => void;
 }
 
-const FilterContainer = ({ handleFilterChange }: FilterContainerProps) => {
+const FilterContainer = ({ filters, handleFilterChange }: FilterContainerProps) => {
   return (
-    <div className={filterLayout}>
-      <ProgressMethodFilter onChange={(value) => handleFilterChange('matchType', value)} />
-      <ContactTargetFilter onChange={handleFilterChange} />
+    <div className={filterContainerLayout}>
+      {/* 진행 방식 필터링 */}
+      <MatchTypeFilter
+        filters={filters}
+        onChange={(value) => handleFilterChange('matchType', value)}
+      />
+
+      {/* 모집 대상 필터링 */}
+      <ContactTargetFilter
+        onChange={handleFilterChange}
+        initialGender={filters.gender}
+        initialAge={filters.age}
+      />
+
+      {/* 지역 필터링 */}
       <AreaFilter onChange={handleFilterChange} />
     </div>
   );
